moved makeswift registrations to makeswift folder
