+++ core/makeswift/components/products-list/products-list.makeswift.tsx
@@ -1,5 +1,6 @@
 import { removeEdgesAndNodes } from '@bigcommerce/catalyst-client';
 import {
+  Checkbox,
   Combobox,
   Image,
   Link,
@@ -14,17 +15,15 @@ import clsx from 'clsx';
 import { useFormatter } from 'next-intl';
 import useSWR from 'swr';
 
+import { CardSkeleton } from '@/vibes/soul/primitives/card';
+import { Price } from '@/vibes/soul/primitives/price-label';
+import { ProductsList } from '@/vibes/soul/primitives/products-list';
 import { SearchProductsResponse } from '~/app/api/products/route';
 import { GetProductsResponse } from '~/client/queries/get-products';
 import { pricesTransformer } from '~/data-transformers/prices-transformer';
 import { runtime } from '~/lib/makeswift/runtime';
 
-import { CardSkeleton } from '../card';
-import { Price } from '../price-label';
-
-import { ProductsCarousel } from '.';
-
-type Product = {
+interface Product {
   entityId?: string;
   title?: string;
   link?: { href?: string; target?: string };
@@ -35,13 +34,15 @@ type Product = {
   type: 'single' | 'range' | 'sale';
   priceOne?: number;
   priceTwo?: number;
-};
+}
 
-type MSProductsCarouselProps = {
+interface MSProductsListProps {
   className: string;
   collection: 'none' | 'bestSelling' | 'newest' | 'featured';
   products: Product[];
-};
+  showCompare: boolean;
+  compareLabel: string;
+}
 
 const fetcher = async (url: string): Promise<{ products: GetProductsResponse }> => {
   const response = await fetch(url);
@@ -55,18 +56,13 @@ const fetcher = async (url: string): Promise<{ products: GetProductsResponse }> 
 };
 
 runtime.registerComponent(
-  function MSProductsCarousel({
-    className,
-    collection,
-    products,
-    ...props
-  }: MSProductsCarouselProps) {
+  function MSProductsList({ className, collection, products, ...props }: MSProductsListProps) {
     const format = useFormatter();
 
     const productIds = products.map(({ entityId }) => entityId ?? '');
 
     const { data, isLoading } = useSWR([collection, productIds], async () => {
-      const apiProducts =
+      const apiResults =
         collection !== 'none'
           ? await fetcher(`/api/products/group/${collection}`)
           : { products: [] };
@@ -77,7 +73,7 @@ runtime.registerComponent(
 
       const additionalProducts = await fetcher(`/api/products/ids?${searchParams.toString()}`);
 
-      return [...apiProducts.products, ...additionalProducts.products];
+      return [...apiResults.products, ...additionalProducts.products];
     });
 
     if (isLoading)
@@ -143,7 +139,7 @@ runtime.registerComponent(
             href: path,
             image: defaultImage ? { src: defaultImage.url, alt: defaultImage.altText } : undefined,
             subtitle: removeEdgesAndNodes(categories)
-              .map(({ name }) => name)
+              .map((category) => category.name)
               .join(', '),
             price: pricesTransformer(prices, format),
           };
@@ -152,14 +148,16 @@ runtime.registerComponent(
 
     const allProducts = [...apiProducts, ...listedProducts];
 
-    return <ProductsCarousel {...props} className={className} products={allProducts} />;
+    return <ProductsList {...props} className={className} products={allProducts} />;
   },
   {
-    type: 'primitive-products-carousel',
-    label: 'Primitives / Products Carousel',
-    icon: 'carousel',
+    type: 'primitive-products-list',
+    label: 'Primitives / Products List',
+    icon: 'gallery',
     props: {
       className: Style(),
+      showCompare: Checkbox({ label: 'Show compare', defaultValue: false }),
+      compareLabel: TextInput({ label: 'Compare label', defaultValue: 'Compare' }),
       collection: Select({
         label: 'Product collection',
         options: [
@@ -171,7 +169,7 @@ runtime.registerComponent(
         defaultValue: 'bestSelling',
       }),
       products: List({
-        label: 'Products',
+        label: 'Additional products',
         type: Shape({
           type: {
             entityId: Combobox({
@@ -211,7 +209,7 @@ runtime.registerComponent(
           },
         }),
         getItemLabel(product) {
-          return product?.title || 'Product';
+          return product?.title || 'Product Title';
         },
       }),
     },
